Use separate state directory for timer script
After a backup is completed, update a state directory outside the backup
directory with the current symlinks and timestamp files.  Update the
timer script to use this state directory instead of the backup directory
to avoid spinning up the backup drives if no backup is needed.

Move the client_lockdir out of the backup directory as well.